+++ src/s21_exp.c
@@ -5,13 +5,13 @@
 long double s21_exp(double x) {
     long double result = 1.0;
     long double sum = 1.0;
-    int i = 1;
+    long double i = 1.0;
     if (x == -S21_INFINITY) return 0;
     if (x == S21_INFINITY) return S21_INFINITY;
     //for (int i = 0; i < 200; i++) {
     while (s21_fabs(result) > S21_EPS) {
         result = result * x / i;
-        i += 1;
+        i += 1.0;
         sum += result; 
         if (sum > DBL_MAX) {
             sum = S21_INFINITY;
